+++ zetaclient/zetacore_observer.go
@@ -5,6 +5,7 @@ import (
 	"math"
 	"time"
 
+	appcontext "github.com/zeta-chain/zetacore/zetaclient/app_context"
 	"github.com/zeta-chain/zetacore/zetaclient/bitcoin"
 	"github.com/zeta-chain/zetacore/zetaclient/interfaces"
 	"github.com/zeta-chain/zetacore/zetaclient/outtxprocessor"
@@ -13,19 +14,14 @@ import (
 
 	sdkmath "cosmossdk.io/math"
 
-	"github.com/pkg/errors"
-	prom "github.com/prometheus/client_golang/prometheus"
 	"github.com/rs/zerolog"
 	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
-	"github.com/zeta-chain/zetacore/zetaclient/config"
 	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 )
 
 const (
-	MaxLookaheadNonce   = 120
-	OutboundTxSignCount = "zetaclient_Outbound_tx_sign_count"
-	HotKeyBurnRate      = "zetaclient_hotkey_burn_rate"
+	MaxLookaheadNonce = 120
 )
 
 type ZetaCoreLog struct {
@@ -38,9 +34,7 @@ type CoreObserver struct {
 	bridge              interfaces.ZetaCoreBridger
 	signerMap           map[common.Chain]interfaces.ChainSigner
 	clientMap           map[common.Chain]interfaces.ChainClient
-	metrics             *metrics.Metrics
 	logger              ZetaCoreLog
-	cfg                 *config.Config
 	ts                  *metrics.TelemetryServer
 	stop                chan struct{}
 	lastOperatorBalance sdkmath.Int
@@ -51,16 +45,13 @@ func NewCoreObserver(
 	bridge interfaces.ZetaCoreBridger,
 	signerMap map[common.Chain]interfaces.ChainSigner,
 	clientMap map[common.Chain]interfaces.ChainClient,
-	metrics *metrics.Metrics,
 	logger zerolog.Logger,
-	cfg *config.Config,
 	ts *metrics.TelemetryServer,
 ) *CoreObserver {
 	co := CoreObserver{
 		ts:   ts,
 		stop: make(chan struct{}),
 	}
-	co.cfg = cfg
 	chainLogger := logger.With().
 		Str("chain", "ZetaChain").
 		Logger()
@@ -73,16 +64,7 @@ func NewCoreObserver(
 	co.signerMap = signerMap
 
 	co.clientMap = clientMap
-	co.metrics = metrics
 	co.logger.ChainLogger.Info().Msg("starting core observer")
-	err := metrics.RegisterCounter(OutboundTxSignCount, "number of Outbound tx signed")
-	if err != nil {
-		co.logger.ChainLogger.Error().Err(err).Msg("error registering counter")
-	}
-	err = metrics.RegisterGauge(HotKeyBurnRate, "Fee burn rate of the hotkey")
-	if err != nil {
-		co.logger.ChainLogger.Error().Err(err).Msg("error registering gauge")
-	}
 	balance, err := bridge.GetZetaHotKeyBalance()
 	if err != nil {
 		co.logger.ChainLogger.Error().Err(err).Msg("error getting last balance of the hot key")
@@ -92,30 +74,10 @@ func NewCoreObserver(
 	return &co
 }
 
-func (co *CoreObserver) Config() *config.Config {
-	return co.cfg
-}
-
-func (co *CoreObserver) GetPromCounter(name string) (prom.Counter, error) {
-	cnt, found := metrics.Counters[name]
-	if !found {
-		return nil, errors.New("counter not found")
-	}
-	return cnt, nil
-}
-
-func (co *CoreObserver) GetPromGauge(name string) (prom.Gauge, error) {
-	gauge, found := metrics.Gauges[name]
-	if !found {
-		return nil, errors.New("gauge not found")
-	}
-	return gauge, nil
-}
-
-func (co *CoreObserver) MonitorCore() {
+func (co *CoreObserver) MonitorCore(appContext *appcontext.AppContext) {
 	myid := co.bridge.GetKeys().GetAddress()
 	co.logger.ZetaChainWatcher.Info().Msgf("Starting Send Scheduler for %s", myid)
-	go co.startCctxScheduler()
+	go co.startCctxScheduler(appContext)
 
 	go func() {
 		// bridge queries UpgradePlan from zetabridge and send to its pause channel if upgrade height is reached
@@ -129,7 +91,7 @@ func (co *CoreObserver) MonitorCore() {
 }
 
 // startCctxScheduler schedules keysigns for cctxs on each ZetaChain block (the ticker)
-func (co *CoreObserver) startCctxScheduler() {
+func (co *CoreObserver) startCctxScheduler(appContext *appcontext.AppContext) {
 	outTxMan := outtxprocessor.NewOutTxProcessorManager(co.logger.ChainLogger)
 	observeTicker := time.NewTicker(3 * time.Second)
 	var lastBlockNum int64
@@ -170,15 +132,10 @@ func (co *CoreObserver) startCctxScheduler() {
 					}
 
 					// Set Current Hot key burn rate
-					gauge, err := co.GetPromGauge(HotKeyBurnRate)
-					if err != nil {
-						co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxEVM: failed to get prometheus gauge: %s for observer", metrics.PendingTxs)
-						continue
-					} // Gauge only takes float values
-					gauge.Set(float64(co.ts.HotKeyBurnRate.GetBurnRate().Int64()))
+					metrics.HotKeyBurnRate.Set(float64(co.ts.HotKeyBurnRate.GetBurnRate().Int64()))
 
 					// schedule keysign for pending cctxs on each chain
-					supportedChains := co.Config().GetEnabledChains()
+					supportedChains := appContext.ZetaCoreContext().GetEnabledChains()
 					for _, c := range supportedChains {
 						if c.ChainId == co.bridge.ZetaChain().ChainId {
 							continue
@@ -190,18 +147,13 @@ func (co *CoreObserver) startCctxScheduler() {
 							co.logger.ZetaChainWatcher.Error().Err(err).Msgf("startCctxScheduler: ListPendingCctx failed for chain %d", c.ChainId)
 							continue
 						}
-						ob, err := co.getUpdatedChainOb(c.ChainId)
+						ob, err := co.getUpdatedChainOb(appContext, c.ChainId)
 						if err != nil {
 							co.logger.ZetaChainWatcher.Error().Err(err).Msgf("startCctxScheduler: getTargetChainOb failed for chain %d", c.ChainId)
 							continue
 						}
 						// Set Pending transactions prometheus gauge
-						gauge, err := ob.GetPromGauge(metrics.PendingTxs)
-						if err != nil {
-							co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxEVM: failed to get prometheus gauge: %s for chain %d", metrics.PendingTxs, c.ChainId)
-							continue
-						}
-						gauge.Set(float64(totalPending))
+						metrics.PendingTxsPerChain.WithLabelValues(c.ChainName.String()).Set(float64(totalPending))
 
 						// #nosec G701 range is verified
 						zetaHeight := uint64(bn)
@@ -257,7 +209,7 @@ func (co *CoreObserver) scheduleCctxEVM(
 		}
 
 		// try confirming the outtx
-		included, _, err := ob.IsSendOutTxProcessed(cctx.Index, params.OutboundTxTssNonce, params.CoinType, co.logger.ZetaChainWatcher)
+		included, _, err := ob.IsSendOutTxProcessed(cctx, co.logger.ZetaChainWatcher)
 		if err != nil {
 			co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxEVM: IsSendOutTxProcessed faild for chain %d nonce %d", chainID, nonce)
 			continue
@@ -337,7 +289,7 @@ func (co *CoreObserver) scheduleCctxBTC(
 			continue
 		}
 		// try confirming the outtx
-		included, confirmed, err := btcClient.IsSendOutTxProcessed(cctx.Index, nonce, params.CoinType, co.logger.ZetaChainWatcher)
+		included, confirmed, err := btcClient.IsSendOutTxProcessed(cctx, co.logger.ZetaChainWatcher)
 		if err != nil {
 			co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxBTC: IsSendOutTxProcessed faild for chain %d nonce %d", chainID, nonce)
 			continue
@@ -365,7 +317,7 @@ func (co *CoreObserver) scheduleCctxBTC(
 	}
 }
 
-func (co *CoreObserver) getUpdatedChainOb(chainID int64) (interfaces.ChainClient, error) {
+func (co *CoreObserver) getUpdatedChainOb(appContext *appcontext.AppContext, chainID int64) (interfaces.ChainClient, error) {
 	chainOb, err := co.getTargetChainOb(chainID)
 	if err != nil {
 		return nil, err
@@ -373,22 +325,23 @@ func (co *CoreObserver) getUpdatedChainOb(chainID int64) (interfaces.ChainClient
 	// update chain client core parameters
 	curParams := chainOb.GetChainParams()
 	if common.IsEVMChain(chainID) {
-		evmCfg, found := co.cfg.GetEVMConfig(chainID)
-		if found && !observertypes.ChainParamsEqual(curParams, evmCfg.ChainParams) {
-			chainOb.SetChainParams(evmCfg.ChainParams)
+		evmParams, found := appContext.ZetaCoreContext().GetEVMChainParams(chainID)
+		if found && !observertypes.ChainParamsEqual(curParams, *evmParams) {
+			chainOb.SetChainParams(*evmParams)
 			co.logger.ZetaChainWatcher.Info().Msgf(
 				"updated chain params for chainID %d, new params: %v",
 				chainID,
-				evmCfg.ChainParams,
+				*evmParams,
 			)
 		}
 	} else if common.IsBitcoinChain(chainID) {
-		_, btcCfg, found := co.cfg.GetBTCConfig()
-		if found && !observertypes.ChainParamsEqual(curParams, btcCfg.ChainParams) {
-			chainOb.SetChainParams(btcCfg.ChainParams)
+		_, btcParams, found := appContext.ZetaCoreContext().GetBTCChainParams()
+
+		if found && !observertypes.ChainParamsEqual(curParams, *btcParams) {
+			chainOb.SetChainParams(*btcParams)
 			co.logger.ZetaChainWatcher.Info().Msgf(
 				"updated chain params for Bitcoin, new params: %v",
-				btcCfg.ChainParams,
+				*btcParams,
 			)
 		}
 	}